tests/text-encoder-polyfill.html

<!DOCTYPE html>
<script src="resources/text-encode-transform.js"></script>
<script src="resources/transform-stream.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

// This test verifies that the polyfill hasn't broken existing functionality.
const inputString = 'I \u{1F499} streams';
const expectedOutputBytes = [73, 32, 240, 159, 146, 153, 32, 115, 116, 114, 101,
                             97, 109, 115];

test(() => {
  const encoder = new TextEncoder();
  assert_array_equals(encoder.encode(inputString), expectedOutputBytes,
                'encoded bytes should match expectation');
}, 'TextEncoder encode() should work');
</script>
